Blood tweaks
